refact/front/navigation: moves backNav.jsx to src/components/navbar/Navbar.jsx
